Allow access via JWT.
This accepts a querystring parameter `token`, consisting of a JWT
containing a subject claim for a fact check id. If the token is valid,
the fact_check_id is passed upstream in the GOVUK_FACT_CHECK_ID header.

Note that this app does not do anything to verify the fact_check_id
itself; that will be done by content-store.

Decoding the token relies on the JWT_AUTH_SECRET env var; the same
value will need to be shared with any publishing app that wants to
generate a token to be used for fact checking.