Add token to the upstream URL
Frontend apps need to be able to access the encrypted web token so
they can generate viable internal URLs.

I can't see a real need to have the token filtered out at the proxy,
so here we add it back in. It's preferable to sharing the encryption
key with all the frontend apps.