Don't meddle with the response transfer headers
We think this is the cause of these two Zendesk tickets:
https://govuk.zendesk.com/agent/tickets/2129122
https://govuk.zendesk.com/agent/tickets/2456890

Unfortunately, the commit message that introduced this
code didn't explain why it was added. I don't think
this is actually true:
>We aren't returning a chunked response

Applications behind the proxy may or may not set this
header and we shouldn't assume they don't. In the case
of Whitehall, it's likely setting this header to
'chunked' when responding with uploaded text file.