Allow iframing draft content
We're working on a new publisher app. Part of that is a page in the
publisher that shows the page in mobile and desktop format. To achieve
this we'll probably have to iframe pages. Currently this is disallowed
by default by Rails. This commit makes sure we strip the
`X-Frame-Options` header so iframing becomes allowed again.

https://trello.com/c/QgznZEUJ