Set auth bypass token into a cookie
This is so that this token can be shared with other apps in the stack,
most notably asset manager. This aims to be the first step to resolving
a problem we have:

A user can view a draft page with a token
That page contains draft assets
Those draft assets can't be viewed because the user is not
authenticated.

It uses a cookie for this approach as it would be impractical to append
query strings to every asset in previewing a document.