Set authenticated user organisation header to be forwarded to content-store
This is to allow access limiting via a users organisation content-id,
this is part of the work publishing workflow are doing to implement
access limiting. Validation will be done in the content-store on whether a user
belonging to a certain organisation can access a document in the draft stack.

Trello:
https://trello.com/c/wxp1XkSg/969-update-govuk-to-support-access-limiting-by-organisation