Proxy both authentication information and JWT token information
Logged in users who also happened to have a JWT token used to not
have any authentication information proxied (only the JWT token
was processed). This meant if the JWT token was invalid or expired,
the user would not be able to access the content, even if they were
otherwise authorised to do so.

This risked causing confusion when we come to implement persisting
tokens as a cookie, as authenticated users would have little idea
that the token is being used rather than their authentication. So
we now always process both.

If no valid JWT token is provided, we call `authenticate!`, which
causes an exception if the user is not logged in. If a valid JWT
token is provided, we downgrade that to `authenticate`, which does
not throw an exception if not logged in.