Stub 'env[warden]' to remove unnecessary conditionals from application
Application checked if 'env[warden]' exists before doing something, but
in reality 'env[warden]' should always exist - we suspect the conditional
is only there because it hadn't been defined/stubbed in the tests. This
fixes that.

In doing so, I had to remove the 'Rack::Lint' unit test added over 4 years
ago in a438e05815e3e38b23895b5116d4d7737bc74cff, as otherwise the test
would fail and complain about the presence of 'non-string in double'.
This implementation is better than the alternative (stubbing methods on
`nil`, and overriding the spec_helper config to allow nil overrides, which
could cause side effects). It appears there is little value in the linter
now, which tries to verify that something looks like a rack application
but actually can't guarantee that and is only getting in the way.