Remove VM hack from application code
Now we're no longer using the VM, we no longer need this workaround.
Besides which, it is not a good idea to change application code just
to hack a test to pass.

This change is required by the next commit, otherwise the domain is
the upstream URI rather than the host as expected.