Authenticate user if not logged in and token is rejected
If you visit a draft page which your token does not allow access to,
rather than showing a "You do not have permission to view this page"
error, we should redirect the user to signon.

This is to prevent issues where a content designer previews a step
by step (which sets the cookie), then indirectly view some other
draft content (without a token in the URL), thereby falling back on
the cookie token which is not compatible with that particular draft
content. Prior to storing tokens as a cookie, the user would be
redirected to signon, so by storing tokens as a cookie, we'll have
caused a regression. This commit keeps the behaviour as it was.