Add note about JWT not being encrpyted
It was news to me that this was not encrypted and the data could actually be decoded and read by anybody. I've not got any evidence that we've used these tokens for sensitive data but feel it's prudent to leave a note for others so they don't then use this to include sensitive information.