Remove issuer documentation; add content_id
This reflects that the guidance on using a user id as an issuer wasn't
representative of the common usage of the field - which is more
commonly used to represent an organisation. Including the user id also
potentially opens up some privacy concerns depending on interpretation
of legislation regarding private ids. The simplest approach is to
remove this value and only consider adding it to guidance when we have
evidence that it is important.

This also updates the definition of the subject to refer to how it
should be unique for an edition of a piece of content, which
differentiates it from some current usages where the same one is used
for all editions.

It also includes the usage of content_id which is suggested to help
identify any tokens that go astray. I'm hoping the combination of
content_id and time can make it possible to track tokens to a request
despite not knowing the user id.

I've made these amends following the incident on 30th June 2020 where
Leicester City Council tweeted a link with an auth_bypass_token.