Run linting as part of default rake task
This applies the convention now used across GOV.UK Ruby apps that
running rake on it's own will perform the necessary test and linting
steps for an app.

It clears the default task if it is already specified (RSpec rails seems
to clobber it) which allows us to redefine it. If we don't do this we
lose the ability to change the order of the task.