Don't proxy RFC 141 healthcheck paths
authenticating-proxy checks that a user is logged in and, if so,
proxies their request to the draft stack.  If the user isn't logged
in, they get redirected to signon to authenticate.

This doesn't work for healthcheck requests, because they need to be
handled by the app.  So there's a check for whether the user is
requesting "/healthcheck".

406106dcb13784a347ab400548b253ac64ee0f9c added a couple more
healthcheck routes, so those also need including in the check.