Add basic Dockerfile to integrate app in GOV.UK ECS
We add a rudimentary Dockerfile to be able to test the integration
of the authentication-proxy app in the draft stack of the new
GOV.UK ECS platform.

It is not intended for this Dockerfile to be used in Production,
this will be done after this [work](https://trello.com/c/GMS8H1bN/426-building-containers-documenting-follow-up)
is completed.

This Dockefile was tested by building the image and testing it with
govuk-infrastructure [PR](https://github.com/alphagov/govuk-infrastructure/pull/251)

Ref:
1. [task trello card](https://trello.com/c/vYNGbJSN/493-add-authenticating-proxy-to-the-draft-stack)