Add initial CI GitHub Action workflow
This workflow enables CI (tests, linting and security checks) using
GitHub Actions. This apart of work to migrate the CI process from
Jenkins to GitHub Actions. This doesn't migrate end to end tests as
they've been deprecated. This also doesn't contain the process for
creating a new release - this remains the responsbility of Jenkins. We
needed to migrate the tests as we should be running tests on merge
commit on the main branch before deploy them to our new infrastructure.