Avoid installing unnecessary libpq-dev package.
govuk-ruby-base already includes libpq5, which is the package that
directly contains the Postgres client shared library.

libpq-dev indirectly installs the shared lib but its main purpose is to
install the headers and so on, which are not needed at runtime. As a
rule of thumb, -dev packages in general should not be necessary at
runtime.

This saves significant time and resources at build-time because we no
longer have to fetch the entire APT catalogue on each build stage.