Add X-Forwarded-Host header
This ensures we pass on the original domain to Rails in the
X-Forwarded-Host, which is used by the redirect_to methods in Rails
frontend applications to generate redirect URLs. This is needed by pages
such as simple smart answers, as Rails redirects user to the next
applicable page. Previously, this was broken because frontend Rails only
had the Host header to generate redirect URLs and this was being set to
the internal domain 'draft-router'.