Remove cookie rotator added for upgrade to Rails 7
Rails 7 changed the default digest class for the key generator from SHA1
to SHA256[1].

The code removed in this commit was added on 15 Mar 2022 in
05a41d8778d6a368a8743163757f91579b0b9103. The intention was to allow any
cookies encrypted using the SHA1 key generator to be read and
re-encrypted using SHA256. I can't see any record of us manually setting
encrypted cookies using `cookies.encrypted`[2] so I think this would
have only applied to our sessions which are stored using the (encrypted)
CookieStore[3]. Given that this rotator was added 17 months ago I think
it's safe to remove it (i.e. I would expect all existing sessions from
17 months ago to have been rotated).

Note that although we don't make any calls to `session` in
authenticating-proxy, there are calls to `session` in the gds-sso Gem
that's included in this app:

- GDS::SSO::FailureApp#store_location![4]
- AuthenticationsController#callback[5]
- lib/gds-sso/warden_config.rb[6]

[1]: https://guides.rubyonrails.org/v7.0/upgrading_ruby_on_rails.html#key-generator-digest-class-changing-to-use-sha256
[2]: https://api.rubyonrails.org/v7.1.1/classes/ActionDispatch/Cookies.html
[3]: https://api.rubyonrails.org/classes/ActionDispatch/Session/CookieStore.html
[4]: https://github.com/alphagov/gds-sso/blob/v18.1.0/lib/gds-sso/failure_app.rb#L48
[5]: https://github.com/alphagov/gds-sso/blob/v18.1.0/app/controllers/authentications_controller.rb#L8
[6]: https://github.com/alphagov/gds-sso/blob/v18.1.0/lib/gds-sso/warden_config.rb#L15