Use RAILS_LOG_TO_STDOUT in production env
Running `rails app:update` introduced this logging configuration which
logs to stdout by default. My guess is that we could probably leave it
as-is but I've added the check for `RAILS_LOG_TO_STDOUT` back in case
we're relying on it to be able to disable logging to stdout in some
production environments.

Note that although the use of `#tap` and `#then` to set the logger were
added by `rails app:update`, the behaviour remains the same.