Use Rails 7.1 default_headers
https://guides.rubyonrails.org/configuring.html#config-action-dispatch-default-headers

This removes the X-Download-Options: "noopen" header which was used to
change the download dialog in Internet Explorer so that users weren't
given the option of opening the file.

The GOV.UK Service Manual[1] says:

> You will still need to support Internet Explorer 11 if your analytics
data shows at least 2% of your users arriving at the service's start
page are using it.

However, the authenticating-proxy app doesn't provide anything to
download so I think it's safe to accept this change without checking
whether we have IE users. Note that the authenticating-proxy might sit
in front of other apps that do allow users to download files, but I'd
expect those apps to retain the IE behaviour if necessary.

[1]: https://www.gov.uk/service-manual/technology/designing-for-different-browsers-and-devices