Use Rails 7.1 log_file_size
https://guides.rubyonrails.org/configuring.html#config-log-file-size

This seems like a sensible change, to limit the size of log files in
test and development environments.

Note that I have to disable RuboCop's `Rails/UnknownEnv` rule as it
doesn't know about `Rails.env.local?`. Presumably because we need to
upgrade rubocop-rails? I'm not going to do anything about this because
the problem will go away once we've switched to the Rails 7.1 defaults
in application.rb and removed this file.