Use Rails 7.1 cache_format_version
https://guides.rubyonrails.org/configuring.html#config-active-support-cache-format-version

We're using the default `:file_store` cache in production[1] so my
expectation is that we're the only app reading from and writing to this
cache store. Because there are no other apps on earlier versions of
Rails sharing this cache store I think it's safe to accept the new
default.

Note that although the information about this change is in
new_framework_defaults_7_1.rb, it explains that the change needs to be
made in application.rb.

[1]: https://guides.rubyonrails.org/configuring.html#config-cache-store