Response headers can now be encoded as Array
In Rack 3 response headers can now be encoded as an Array when
multiple cookies are set[1]. Previously this test assumed that they
were encoded as a string with separate cookies separated by "\n". To
maintain the intention of the test (checking that the domain of the
auth_bypass_token cookie is set correctly) we cast the headers to an
Array (as a single cookie is still returned as a string), find the
cookie with the `auth_bypass_token=` string, and then check the domain
matches the expected domain.

[1] https://github.com/rack/rack/blob/main/UPGRADE-GUIDE.md#multiple-response-header-values-are-encoded-using-an-array